refactor: move cnudie.util helpers into ocm, version modules
cnudie.util became a thin re-export shim (similar to cnudie.retrieve).
Functions relocated to:

- ocm/__init__.py: ComponentId/ComponentName aliases (widened),
  to_component_id, to_component_name, to_component, iter_sorted
- ocm/diff.py (new): ComponentDiff, LabelDiff, ResourceDiff,
  ComponentResource, diff_components, diff_labels, diff_resources,
  format_component_diff
- ocm/access.py: oci_ref, target_oci_ref, oci_artefact_reference,
  normalise_component_name, to_component_id_and_repository_url
- ocm/util.py: main_source gained label heuristic
  (cloud.gardener/cicd/source, repository-classification=main)
- version.py: META_SEPARATOR, sanitise_version, desanitise_version

Dropped: determine_component_name (no callers).

Also:
- ocm/retrieve{,_async}.py bootstrap off cnudie.util
- github/pullrequest.py uses ocm.diff
- test/cnudie/cnudie_util_test.py -> test/ocm/ocm_util_test.py,
  retargeted at new homes
