+++ .github/actions/component-diff/action.yaml
@@ -2,7 +2,7 @@ name: component-diff
 description: |
   An action to determine the component-diff between the current version (read from passed-in
   component-descriptor) and the greatest published release-version. The diff is structured according
-  to `cnudie.util.ComponentDiff`.
+  to `ocm.diff.ComponentDiff`.
 
 inputs:
   component-descriptor:
