+++ test/ocm/ocm_util_test.py
@@ -2,15 +2,11 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-
 import pytest
 
-import cnudie.util
 import ocm
-
-# functions under test
-diff_components = cnudie.util.diff_components
-diff_resources = cnudie.util.diff_resources
+import ocm.diff
+import ocm.util
 
 
 @pytest.fixture
@@ -65,17 +61,17 @@ def iref():
 
 
 def test_componentdiff_to_str(cid):
-    diff = cnudie.util.ComponentDiff(
+    diff = ocm.diff.ComponentDiff(
         cidentities_only_left=(cid(name='a', version='1.0.0'), cid(name='b', version='2.0.0')),
         cidentities_only_right=(cid(name='c', version='3.0.0'),),  # added
         cpairs_version_changed=[
-            (comp('d', '1.0.0'), comp('d', '2.0.0'))  # changed
+            (comp('d', '1.0.0'), comp('d', '2.0.0')),  # changed
         ],
         names_only_left={'a', 'b'},
         names_only_right={'c'},
         names_version_changed={'d'},
     )
-    formatted_diff = cnudie.util.format_component_diff(
+    formatted_diff = ocm.diff.format_component_diff(
         component_diff=diff,
         delivery_dashboard_url_view_diff=None,
         delivery_dashboard_url=None,
@@ -111,7 +107,7 @@ def test_iter_sorted():
     comp_b = comp(name='b', version=1, componentReferences=[cref(comp_a)])
     comp_c = comp(name='c', version=1, componentReferences=[cref(comp_a), cref(comp_b)])
 
-    sorted_comps = tuple(cnudie.util.iter_sorted((comp_c, comp_b, comp_a)))
+    sorted_comps = tuple(ocm.iter_sorted((comp_c, comp_b, comp_a)))
 
     assert sorted_comps == (comp_a, comp_b, comp_c)
 
@@ -123,12 +119,15 @@ def test_remove_component(cid):
     ]
 
     right_components = [
-        comp('c1', '1.2.3'), # version changed
-        # comp('c2', '1.2.3'), # removed
+        comp('c1', '1.2.3'),  # version changed
+        # comp('c2', '1.2.3'),  # removed
     ]
 
-    result = diff_components(left_components=left_components, right_components=right_components)
-    assert result.cidentities_only_left ==  {cid('c2', 'v2.0.0')}
+    result = ocm.diff.diff_components(
+        left_components=left_components,
+        right_components=right_components,
+    )
+    assert result.cidentities_only_left == {cid('c2', 'v2.0.0')}
     assert result.cidentities_only_right == set()
     assert result.cpairs_version_changed == []
     assert result.names_only_left == {'c2'}
@@ -139,8 +138,11 @@ def test_remove_component(cid):
     left_components.append(comp('c2', 'v1.4.0'))
     left_components.append(comp('c2', 'v1.5.0'))
 
-    result = diff_components(left_components=left_components, right_components=right_components)
-    assert result.cidentities_only_left ==  {
+    result = ocm.diff.diff_components(
+        left_components=left_components,
+        right_components=right_components,
+    )
+    assert result.cidentities_only_left == {
         cid('c2', 'v1.4.0'),
         cid('c2', 'v1.5.0'),
     }
@@ -156,19 +158,19 @@ def test_diff_components(cid):
         comp('c1', '1.2.3'),
         comp('c2', '1.2.3'),
         comp('c3', '1.2.3'),
-        # comp('c4', '1.2.3'), # missing on left
-        comp('c5', '1.2.3'), # version changed
+        # comp('c4', '1.2.3'),  # missing on left
+        comp('c5', '1.2.3'),  # version changed
     )
 
     right_components = (
-        comp('c1', '2.2.3'), # version changed
-        comp('c2', '1.2.3'), # no change
-        #cid('c3', '1.2.3'), # missing on right
-        comp('c4', '1.2.3'), # added on right
-        comp('c5', '2.3.4'), # version changed
+        comp('c1', '2.2.3'),  # version changed
+        comp('c2', '1.2.3'),  # no change
+        # cid('c3', '1.2.3'),  # missing on right
+        comp('c4', '1.2.3'),  # added on right
+        comp('c5', '2.3.4'),  # version changed
     )
 
-    result = diff_components(left_components, right_components)
+    result = ocm.diff.diff_components(left_components, right_components)
 
     assert result.cidentities_only_left == {
         cid('c1', '1.2.3'), cid('c3', '1.2.3'), cid('c5', '1.2.3'),
@@ -182,10 +184,10 @@ def test_diff_components(cid):
     ]
     assert result.names_only_left == {'c3'}
     assert result.names_only_right == {'c4'}
-    assert result.names_version_changed == {'c1','c5'}
+    assert result.names_version_changed == {'c1', 'c5'}
 
 
-#TODO add other resources than OCI images
+# TODO add other resources than OCI images
 def test_diff_resources(iref):
     left_comp = comp('x.o/a/b', '1.2.3')
     right_comp = comp('x.o/a/b', '2.3.4')
@@ -195,9 +197,9 @@ def test_diff_resources(iref):
     left_comp.resources.append(img1)
     right_comp.resources.append(img1)
 
-    img_diff = diff_resources(left_component=left_comp, right_component=right_comp)
+    img_diff = ocm.diff.diff_resources(left_component=left_comp, right_component=right_comp)
 
-    # same image added declared by left and right - expect empty diff
+    # same image declared by left and right - expect empty diff
     assert img_diff.left_component == left_comp
     assert img_diff.right_component == right_comp
     assert len(img_diff.resource_refs_only_right) == 0
@@ -209,17 +211,17 @@ def test_diff_resources(iref):
     right_comp.resources.append(img3)
 
     # img2 only left, img3 only right
-    resource_diff = diff_resources(left_component=left_comp, right_component=right_comp)
+    resource_diff = ocm.diff.diff_resources(left_component=left_comp, right_component=right_comp)
     assert len(resource_diff.resource_refs_only_left) == 1
     assert len(resource_diff.resource_refs_only_right) == 1
     assert list(resource_diff.resource_refs_only_left)[0] == img2
     assert list(resource_diff.resource_refs_only_right)[0] == img3
 
     img4_0 = iref('i4', '1.2.3')
-    img4_1 = iref('i4', '2.0.0') # changed version
+    img4_1 = iref('i4', '2.0.0')  # changed version
     left_comp.resources.append(img4_0)
     right_comp.resources.append(img4_1)
-    resource_diff = diff_resources(left_component=left_comp, right_component=right_comp)
+    resource_diff = ocm.diff.diff_resources(left_component=left_comp, right_component=right_comp)
     assert len(resource_diff.resource_refs_only_left) == 1
     assert len(resource_diff.resource_refs_only_right) == 1
     assert len(resource_diff.resourcepairs_version_changed) == 1
@@ -235,18 +237,18 @@ def test_diff_resources(iref):
     left_comp.resources.append(img5_0)
     left_comp.resources.append(img5_1)
     left_comp.resources.append(img5_2)
-    resource_diff = diff_resources(left_component=left_comp, right_component=right_comp)
+    resource_diff = ocm.diff.diff_resources(left_component=left_comp, right_component=right_comp)
     assert len(resource_diff.resource_refs_only_left) == 4
     assert len(resource_diff.resource_refs_only_right) == 1
     assert list(resource_diff.resource_refs_only_left)[1] == img5_0
     assert list(resource_diff.resource_refs_only_left)[2] == img5_1
     assert list(resource_diff.resource_refs_only_left)[3] == img5_2
 
-    # test if grouping semantic does work
+    # test if grouping semantics work
     right_comp.resources.append(img5_0)
     img5_3 = iref('res5', '1.2.6')
     right_comp.resources.append(img5_3)
-    resource_diff = diff_resources(left_component=left_comp, right_component=right_comp)
+    resource_diff = ocm.diff.diff_resources(left_component=left_comp, right_component=right_comp)
 
     assert len(resource_diff.resource_refs_only_left) == 2
     assert len(resource_diff.resource_refs_only_right) == 1
@@ -262,7 +264,7 @@ def test_label_usage():
             access=ocm.GithubAccess(
                 type=ocm.AccessType.GITHUB,
                 ref='refs/heads/master',
-                repoUrl='github.com/otherOrg/otherRepo'
+                repoUrl='github.com/otherOrg/otherRepo',
             ),
             labels=[
                 ocm.Label(
@@ -276,7 +278,7 @@ def test_label_usage():
             access=ocm.GithubAccess(
                 type=ocm.AccessType.GITHUB,
                 ref='refs/heads/master',
-                repoUrl='github.com/org/repo'
+                repoUrl='github.com/org/repo',
             ),
             labels=[
                 ocm.Label(
@@ -302,7 +304,7 @@ def test_label_usage():
         provider=[],
     )
 
-    main_source = cnudie.util.determine_main_source_for_component(component_with_source_label,)
+    main_source = ocm.util.main_source(component_with_source_label)
     assert main_source.labels[0].value == {'repository-classification': 'main'}
     assert main_source.name == 'repo_main_source'
 
@@ -315,7 +317,7 @@ def test_label_usage():
                 access=ocm.GithubAccess(
                     type=ocm.AccessType.GITHUB,
                     ref='refs/heads/master',
-                    repoUrl='github.com/org/repo'
+                    repoUrl='github.com/org/repo',
                 ),
             ),
             ocm.Source(
@@ -323,7 +325,7 @@ def test_label_usage():
                 access=ocm.GithubAccess(
                     type=ocm.AccessType.GITHUB,
                     ref='refs/heads/master',
-                    repoUrl='github.com/otherOrg/otherRepo'
+                    repoUrl='github.com/otherOrg/otherRepo',
                 ),
             ),
         ],
@@ -339,7 +341,7 @@ def test_label_usage():
         provider=[],
     )
 
-    main_source = cnudie.util.determine_main_source_for_component(component_without_source_label)
+    main_source = ocm.util.main_source(component_without_source_label)
 
     assert main_source.name == 'repo_main_source'
 
@@ -347,66 +349,62 @@ def test_label_usage():
 def test_diff_label():
     label_foo = ocm.Label(name='foo', value='bar v1')
 
-    left_labels = [
-        label_foo
-    ]
-    right_labels = [
-        label_foo
-    ]
+    left_labels = [label_foo]
+    right_labels = [label_foo]
 
-    # check identical label in both lists
-    label_diff = cnudie.util.diff_labels(left_labels=left_labels, right_labels=right_labels)
+    # identical labels in both lists
+    label_diff = ocm.diff.diff_labels(left_labels=left_labels, right_labels=right_labels)
     assert len(label_diff.label_pairs_changed) == 0
     assert len(label_diff.labels_only_left) == 0
     assert len(label_diff.labels_only_right) == 0
 
-    # check left exclusive label
+    # left-exclusive label
     label_only_left = ocm.Label(name='left', value='only')
     left_labels.append(label_only_left)
-    label_diff = cnudie.util.diff_labels(left_labels=left_labels, right_labels=right_labels)
+    label_diff = ocm.diff.diff_labels(left_labels=left_labels, right_labels=right_labels)
     assert len(label_diff.label_pairs_changed) == 0
     assert len(label_diff.labels_only_left) == 1
     assert label_diff.labels_only_left[0] == label_only_left
     assert len(label_diff.labels_only_right) == 0
 
-    # check right exclusive label
+    # right-exclusive label
     label_only_right = ocm.Label(name='right', value='only')
     right_labels.append(label_only_right)
-    label_diff = cnudie.util.diff_labels(left_labels=left_labels, right_labels=right_labels)
+    label_diff = ocm.diff.diff_labels(left_labels=left_labels, right_labels=right_labels)
     assert len(label_diff.label_pairs_changed) == 0
     assert len(label_diff.labels_only_left) == 1
     assert len(label_diff.labels_only_right) == 1
     assert label_diff.labels_only_right[0] == label_only_right
 
-    # check removal of one label
+    # removal of one label
     right_labels.remove(label_foo)
-    label_diff = cnudie.util.diff_labels(left_labels=left_labels, right_labels=right_labels)
+    label_diff = ocm.diff.diff_labels(left_labels=left_labels, right_labels=right_labels)
     assert len(label_diff.label_pairs_changed) == 0
     assert len(label_diff.labels_only_left) == 2
     assert label_diff.labels_only_left[0] == label_foo
     assert len(label_diff.labels_only_right) == 1
 
-    # check different label value with the same name
+    # same name, different value
     label_foo_updated = ocm.Label(name='foo', value='bar v2')
     right_labels.append(label_foo_updated)
-    label_diff = cnudie.util.diff_labels(left_labels=left_labels, right_labels=right_labels)
+    label_diff = ocm.diff.diff_labels(left_labels=left_labels, right_labels=right_labels)
     assert len(label_diff.label_pairs_changed) == 1
     assert label_diff.label_pairs_changed[0] == (label_foo, label_foo_updated)
     assert len(label_diff.labels_only_left) == 1
     assert len(label_diff.labels_only_right) == 1
 
-    # check that duplicate label name in one list cause exception
+    # duplicate label names in one list must raise
     right_labels.append(label_foo_updated)
     with pytest.raises(RuntimeError) as re:
-        label_diff = cnudie.util.diff_labels(left_labels=left_labels, right_labels=right_labels)
-    assert re != None
+        ocm.diff.diff_labels(left_labels=left_labels, right_labels=right_labels)
+    assert re is not None
 
 
 def test_to_component_id():
     base_identity = ocm.ComponentIdentity(name='Foo', version='1.2.3')
 
     test_identity = ocm.ComponentIdentity(name='Foo', version='1.2.3')
-    assert cnudie.util.to_component_id(test_identity) == base_identity
+    assert ocm.to_component_id(test_identity) == base_identity
 
     test_component = ocm.Component(
         name='Foo',
@@ -417,36 +415,36 @@ def test_to_component_id():
         componentReferences=[],
         resources=[],
     )
-    assert cnudie.util.to_component_id(test_component) == base_identity
+    assert ocm.to_component_id(test_component) == base_identity
 
     test_component_descriptor = ocm.ComponentDescriptor(
         meta=ocm.Metadata(),
         component=test_component,
         signatures=[],
     )
-    assert cnudie.util.to_component_id(test_component_descriptor) == base_identity
+    assert ocm.to_component_id(test_component_descriptor) == base_identity
 
     test_component_reference = ocm.ComponentReference(
         componentName='Foo', name='Bar', version='1.2.3',
     )
-    assert cnudie.util.to_component_id(test_component_reference) == base_identity
+    assert ocm.to_component_id(test_component_reference) == base_identity
 
     test_str = 'Foo:1.2.3'
-    assert cnudie.util.to_component_id(test_str) == base_identity
+    assert ocm.to_component_id(test_str) == base_identity
 
     test_str = 'Foo'
     with pytest.raises(ValueError):
-        cnudie.util.to_component_id(test_str)
+        ocm.to_component_id(test_str)
 
     test_tuple = 'Foo', '1.2.3'
-    assert cnudie.util.to_component_id(test_tuple) == base_identity
+    assert ocm.to_component_id(test_tuple) == base_identity
 
 
 def test_to_component_name():
     base_name = 'Foo'
 
     test_identity = ocm.ComponentIdentity(name='Foo', version='1.2.3')
-    assert cnudie.util.to_component_name(test_identity) == base_name
+    assert ocm.to_component_name(test_identity) == base_name
 
     test_component = ocm.Component(
         name='Foo',
@@ -457,26 +455,26 @@ def test_to_component_name():
         componentReferences=[],
         resources=[],
     )
-    assert cnudie.util.to_component_name(test_component) == base_name
+    assert ocm.to_component_name(test_component) == base_name
 
     test_component_descriptor = ocm.ComponentDescriptor(
         meta=ocm.Metadata(),
         component=test_component,
         signatures=[],
     )
-    assert cnudie.util.to_component_name(test_component_descriptor) == base_name
+    assert ocm.to_component_name(test_component_descriptor) == base_name
 
     test_component_reference = ocm.ComponentReference(
         componentName='Foo', name='Bar', version='1.2.3',
     )
-    assert cnudie.util.to_component_name(test_component_reference) == base_name
+    assert ocm.to_component_name(test_component_reference) == base_name
 
     test_str = 'Foo:1.2.3'
-    assert cnudie.util.to_component_name(test_str) == base_name
+    assert ocm.to_component_name(test_str) == base_name
 
     test_str = 'Foo:Bar:Baz'
     with pytest.raises(ValueError):
-        cnudie.util.to_component_name(test_str)
+        ocm.to_component_name(test_str)
 
     test_tuple = 'Foo', '1.2.3'
-    assert cnudie.util.to_component_name(test_tuple) == base_name
+    assert ocm.to_component_name(test_tuple) == base_name
